Added lightmap test.
This version uses manually optimized scene and models exported from Unity.

Main difference to current automatic exports are that "real" materials and textures are now defined in the scene, not in models. This allow loader both to reuse lightmap texture and to synchronize on loading of textures.

Also folder structure is bit more ordered ;)

Shouldn't be too difficult to modify Unity exporter to create this automatically.