Even though changes are made in test/sound these changes have very little to do with the sound (a bit, though). I've updated:
Matrix4.lookAt() : now it sets the rotation-part of the matrix as it should (before it was transposed, kind of). Also, it sets the position to eye.xyz
WebGLRenderer() : line 2303 : changed so it's using the camera.inverseMatrix instead of .globalMatrix, as it's supposed to be.
WebGLRenderer() : line 1753 : forced isInFrustum to return true, as the incoming matrix to computeFrustum isn't working any more. The frustum check I wrote might be working now - please try it :)
SoundRenderer: multiplying the Sound3D.globalMatrix with the camera.inverseMatrix as it's supposed to be.

Probably the skin shader needs to be updated/reverted to the original implementation.