Made workaround hack for Chrome crashing in viewer 3.
Seems to be something threading / timing related: it doesn't crash when debugging or when popping alerts.

Which inspired me to call image drop handler from setTimeout. This seems to work, but needs quite some delay not to crash. Fortunately delay is not very noticeable even with 500ms per image. Seems like all files from one drop are handled in parallel (which may be the source of threading issue).