Model viewer: replaced missing GUI-DAT sources links with minified one. Fixed broken antialiasing with transparent background.
GUI-DAT still looks different than in the demo on Google network, I guess it's because of missing custom CSS?

For antialiasing: it make troubles on Chrome ANGLE if used with transparent background of the WebGL canvas DOM element, creating slight outline. It's better just to use opaque clear color matching the background color of the document.